Refactor: Splitting the password generation module into its utilities in different files, adding a recoder module for when salt is known
